renamed function and changed structure
